+++ src/Markdown.MAML/Transformer/ModelTransformerBase.cs
@@ -24,8 +24,6 @@ namespace Markdown.MAML.Transformer
         internal const int EXAMPLE_HEADING_LEVEL = 3;
         internal const int PARAMETERSET_NAME_HEADING_LEVEL = 3;
 
-        private static Regex DetectPSLanguageExpression = new Regex(@"^(PS C:\\>| {0,}[a-z]{3,11}-[a-z0-9]{2,}|#)", RegexOptions.IgnoreCase | RegexOptions.Compiled);
-
         public ModelTransformerBase(Action<string> infoCallback, Action<string> warningCallback)
         {
             _infoCallback = infoCallback;
@@ -187,7 +185,7 @@ namespace Markdown.MAML.Transformer
                 {
                     codeBlocks.Add(new MamlCodeBlock(
                         codeBlockNode.Text,
-                        string.IsNullOrEmpty(codeBlockNode.LanguageMoniker) ? DetectLanguage(codeBlockNode.Text) : codeBlockNode.LanguageMoniker
+                        codeBlockNode.LanguageMoniker
                     ));
                 }
 
@@ -204,20 +202,6 @@ namespace Markdown.MAML.Transformer
                 throw headingException;
             }
             
-        }
-
-        private string DetectLanguage(string hint)
-        {
-            // Detect PowerShell based on first line
-            // - Look PS C:\> prefix followed 
-            // - Look for standalone PowerShell verbs-noun, current verbs use 3-11 characters
-            // - Look for inline help # character
-            if (DetectPSLanguageExpression.IsMatch(hint))
-            {
-                return "powershell";
-            }
-
-            return string.Empty;
         }
 
         protected void RelatedLinksRule(MamlCommand commmand)