+++ test/Pester/PlatyPs.Tests.ps1
@@ -260,11 +260,11 @@ Describe 'New-MarkdownHelp' {
             System.String. Add-Extension returns a string with the extension or file name.
 
             .EXAMPLE
-            C:\PS> Test-PlatyPSFunction "File"
+            PS C:\> Test-PlatyPSFunction "File"
             File.txt
 
             .EXAMPLE
-            C:\PS> Test-PlatyPSFunction "File" -First "doc"
+            PS C:\> Test-PlatyPSFunction "File" -First "doc"
             File.doc
 
             .LINK
@@ -306,6 +306,10 @@ Describe 'New-MarkdownHelp' {
 
         It 'generates markdown with placeholder for parameter with no description' {
             ($content | Where-Object {$_ -eq '{{Fill Common Description}}'} | Measure-Object).Count | Should Be 1
+        }
+
+        It 'generates markdown example powershell code blocks' {
+            $content | Out-String | Should Match '(\w|\W)+\r\n### EXAMPLE 1\r\n```powershell\r\nTest-PlatyPSFunction "File"\r\n```'
         }
     }